chore move ShareSliceContent codec from test to core
