+++ packages/core/src/libraries/mocks.ts
@@ -1,8 +1,8 @@
 import path from "path";
 import { isRight } from "fp-ts/Either";
-
+import * as t from "io-ts";
 import { Files } from "../node-utils";
-import { SliceMock } from "../models";
+import { SliceMock, SharedSliceContent } from "../models";
 
 export function createPathToMock({
   path: filePath,
@@ -24,16 +24,18 @@ export function resolvePathsToMock({
   paths: ReadonlyArray<string>;
   from: string;
   sliceName: string;
-}): { path: string; value: SliceMock | undefined } | undefined {
+}):
+  | { path: string; value: SliceMock | SharedSliceContent | undefined }
+  | undefined {
   const possiblePaths = paths.map((base) =>
     createPathToMock({ path: base, from, sliceName })
   );
-  return Files.readFirstOf<SliceMock | undefined>(possiblePaths)(
-    (v: string) => {
-      const res = SliceMock.decode(JSON.parse(v));
-      if (isRight(res)) {
-        return res.right;
-      }
+  return Files.readFirstOf<SliceMock | SharedSliceContent | undefined>(
+    possiblePaths
+  )((v: string) => {
+    const res = t.union([SliceMock, SharedSliceContent]).decode(JSON.parse(v)); // here
+    if (isRight(res)) {
+      return res.right;
     }
-  );
+  });
 }
