+++ packages/core/src/models/Slice.ts
@@ -7,6 +7,29 @@ import {
 } from "@prismicio/types-internal/lib/customtypes/widgets/slices";
 import { getOrElseW } from "fp-ts/lib/Either";
 import { FieldsSM } from "./Fields";
+import { FieldContentType } from "@prismicio/types-internal/lib/documents/widgets/nestable/FieldContent";
+import {
+  EmptyContent,
+  EmptyContentType,
+  GroupItemContentType,
+  UIDContent,
+  UIDContentType,
+} from "@prismicio/types-internal/lib/documents/widgets";
+import {
+  BooleanContent,
+  EmbedContent,
+  GeoPointContent,
+  ImageContent,
+  IntegrationFieldsContent,
+  type SeparatorContent,
+  SeparatorContentType,
+  StructuredTextContent,
+  StructuredTextContentType,
+} from "@prismicio/types-internal/lib/documents/widgets/nestable";
+import { BooleanContentType } from "@prismicio/types-internal/lib/documents/widgets/nestable/BooleanContent";
+import { LinkContent } from "@prismicio/types-internal/lib/documents/widgets/nestable/Link";
+import { LinkContentType } from "@prismicio/types-internal/lib/documents/widgets/nestable/Link/LinkContent";
+import { SharedSliceContentType } from "@prismicio/types-internal/lib/documents/widgets/slices";
 
 const IMAGE_PLACEHOLDER_URL =
   "https://images.prismic.io/slice-machine/621a5ec4-0387-4bc5-9860-2dd46cbc07cd_default_ss.png?auto=compress,format";
@@ -127,3 +150,77 @@ export const Slices = {
     );
   },
 };
+
+const SeparatorContentC: t.Type<SeparatorContent> = t.type({
+  __TYPE__: t.literal(SeparatorContentType),
+});
+
+const StructuredTextContentC = t.type({
+  __TYPE__: t.literal(StructuredTextContentType),
+  value: StructuredTextContent,
+});
+
+const FieldTypes: Record<string, null> = [
+  "Text",
+  "Date",
+  "Timestamp",
+  "Color",
+  "Number",
+  "Range",
+  "Select",
+].reduce((acc, curr) => {
+  return { ...acc, [curr]: null };
+}, {}); // this causes some issues
+
+const FieldContentC = t.type({
+  __TYPE__: t.literal(FieldContentType),
+  type: t.keyof(FieldTypes),
+  value: t.string,
+});
+
+const EmptyContentC: t.Type<EmptyContent> = t.type({
+  __TYPE__: t.literal(EmptyContentType),
+  type: t.string,
+});
+
+const SimpleWidgetContent /* t.Type<SimpleWidgetContentT> */ = t.union([
+  IntegrationFieldsContent,
+  StructuredTextContentC,
+  ImageContent,
+  t.type({
+    __TYPE__: t.literal(GroupItemContentType),
+    value: GeoPointContent,
+  }),
+  GeoPointContent, // weird that geo has no __TYPE__
+  EmbedContent,
+  t.type({
+    __TYPE__: t.literal(LinkContentType),
+    value: LinkContent,
+  }),
+  t.type({
+    __TYPE__: t.literal(UIDContentType),
+    value: UIDContent,
+  }),
+  SeparatorContentC,
+  FieldContentC,
+  t.type({
+    __TYPE__: t.literal(BooleanContentType),
+    value: BooleanContent,
+  }),
+  EmptyContentC,
+]);
+
+const GroupItemContent /*: t.Type<GroupItemContentT> */ = t.type({
+  __TYPE__: t.literal(GroupItemContentType),
+  value: t.array(t.tuple([t.string, SimpleWidgetContent])),
+});
+
+const SharedSliceContentItem /*: t.Type<SharedSliceContentT> */ = t.type({
+  variation: t.string,
+  primary: t.record(t.string, SimpleWidgetContent),
+  items: t.array(GroupItemContent),
+  __TYPE__: t.literal(SharedSliceContentType),
+});
+
+export const SharedSliceContent = t.array(SharedSliceContentItem);
+export type SharedSliceContent = t.TypeOf<typeof SharedSliceContent>;
